refactor(earn): dvv and ggv components for improved availability checks and warnings
