+++ features/earn/vault-dvv/hooks/use-dvv-available.ts
