+++ features/earn/vault-dvv/withdraw/dvv-withdraw-submit.tsx
@@ -1,5 +1,5 @@
 import { VaultSubmitButton } from 'features/earn/shared/vault-submit-button';
-import { useDVVAvailable } from '../hooks/use-dvv-avaliable';
+import { useDVVAvailable } from '../hooks/use-dvv-available';
 
 export const DVVWithdrawSubmitButton = () => {
   const { isDVVAvailable } = useDVVAvailable();
