+++ features/earn/vault-ggv/deposit/ggv-deposit-form.tsx
@@ -6,7 +6,7 @@ import { VaultTxInfo } from 'features/earn/shared/vault-tx-info';
 
 import { GGVDepositFormProvider } from './form-context';
 import { GGVDepositInputGroup } from './ggv-deposit-input-group';
-import { GGVAvailableDeposit } from './ggv-avaliable-deposit';
+import { GGVAvailableDeposit } from './ggv-available-deposit';
 import { GGVWillReceive } from './ggv-deposit-will-receive';
 import { GGVDepositSubmitButton } from './ggv-deposit-submit-botton';
 import { GGVDepositWarning } from './ggv-deposit-warning';
